requirements.txt
celery==5.4.0
crispy-bootstrap5==2024.2
Django==5.0.7
django-crispy-forms==2.3
djangorestframework==3.15.2
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
gunicorn==23.0.0
django-environ==0.11.2